+++ internal/ansible/ansible.go
@@ -14,7 +14,7 @@ import (
 	"github.com/gofastercloud/boetticher/internal/firewall"
 	"github.com/gofastercloud/boetticher/internal/logging"
 	"github.com/gofastercloud/boetticher/internal/model"
-	"github.com/gofastercloud/boetticher/internal/zabbix"
+	"github.com/gofastercloud/boetticher/internal/pulse"
 )
 
 func Inventory(s model.Site) (string, error) {
@@ -118,7 +118,7 @@ func Variables(s model.Site) ([]byte, error) {
 	if err != nil {
 		return nil, err
 	}
-	zabbixPlan, err := zabbix.PlanFromSite(s)
+	monitoringPlan, err := pulse.PlanFromSite(s)
 	if err != nil {
 		return nil, err
 	}
@@ -155,7 +155,7 @@ func Variables(s model.Site) ([]byte, error) {
 		AdGuardForwardZones         []string                      `json:"adguard_forward_zones"`
 		DNSPlan                     dns.Plan                      `json:"dns_plan"`
 		FirewallPlan                firewall.Plan                 `json:"firewall_plan"`
-		ZabbixPlan                  zabbix.Plan                   `json:"zabbix_plan"`
+		MonitoringPlan              pulse.Plan                    `json:"monitoring_plan"`
 		BlockyConfig                string                        `json:"blocky_config"`
 		LoggingPlan                 logging.Plan                  `json:"logging_plan"`
 		LoggingCollectorConfig      string                        `json:"logging_collector_config"`
@@ -165,7 +165,7 @@ func Variables(s model.Site) ([]byte, error) {
 		LoggingCollectorCertificate string                        `json:"logging_collector_certificate"`
 		ModuleConfigs               map[string]model.ModuleConfig `json:"module_configs"`
 		ModuleDeclarations          []model.ModuleDeclaration     `json:"module_declarations"`
-	}{revision, s.Network.Domain, true, dnsPlan.Implementation, dnsPlan.ImplementationVersion, dnsPlan.PackageVersion, dns.AuthoritativePort, dynamicZoneNames(dnsPlan.DynamicZones), dnsPlan.AdGuardForwardZones, dnsPlan, firewallPlan, zabbixPlan, string(blockyConfig), loggingPlan, logging.CollectorConfiguration(loggingPlan), logging.CollectorServiceOverride(loggingPlan), loggingUploads, map[string]string{}, "", s.ModuleConfig, s.Declarations}
+	}{revision, s.Network.Domain, true, dnsPlan.Implementation, dnsPlan.ImplementationVersion, dnsPlan.PackageVersion, dns.AuthoritativePort, dynamicZoneNames(dnsPlan.DynamicZones), dnsPlan.AdGuardForwardZones, dnsPlan, firewallPlan, monitoringPlan, string(blockyConfig), loggingPlan, logging.CollectorConfiguration(loggingPlan), logging.CollectorServiceOverride(loggingPlan), loggingUploads, map[string]string{}, "", s.ModuleConfig, s.Declarations}
 	data, err := json.MarshalIndent(value, "", "  ")
 	if err != nil {
 		return nil, err